implement basic parse
